fix: renaming module
